# utils/swagger/vehicles.yaml
openapi: 3.0.0
info:
  title: Vehicle Management API
  version: 1.0.0
  description: API for managing vehicles, services, tyres, and gallery
...
paths:
...
  /vehicles:
    post:
      tags: [Vehicles]
      summary: Create a new vehicle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - user_id
                - vehicle_type_id
                - make
                - title
              properties:
                user_id:
                  type: integer
                  example: 1
                country_id:
                  type: integer
                  example: 1
                state_id:
                  type: integer
                  example: 1
                city_id:
                  type: integer
                  example: 1
                vehicle_type_id:
                  type: integer
                  example: 1
                make:
                  type: string
                  example: "Toyota"
                title:
                  type: string
                  example: "Toyota Camry 2023"
                model:
                  type: string
                  example: "Camry"
                year_manufactured:
                  type: integer
                  example: 2023
                color:
                  type: string
                  example: "White"
                trim:
                  type: string
                  example: "LE"
                fuel_type:
                  type: string
                  enum: [Petrol, Diesel, CNG, Electric, Hybrid, Other]
                  example: "Petrol"
                registration_number:
                  type: string
                  example: "ABC123"
                chassis_number:
                  type: string
                  example: "1HGCM82633A123456"
                base_location:
                  type: string
                  example: "New York"
                registration_expiry_date:
                  type: string
                  format: date
                  example: "2025-12-31"
                sitting_capacity:
                  type: integer
                  example: 5
                luggage_capacity:
                  type: integer
                  example: 2
                speed_limit:
                  type: integer
                  example: 120
                purchase_date:
                  type: string
                  format: date
                  example: "2023-01-15"
                purchase_amount:
                  type: number
                  format: float
                  example: 25000.00
                loan_price:
                  type: number
                  format: float
                  example: 20000.00
                number_of_emi_months:
                  type: integer
                  example: 60
                emi_amount:
                  type: number
                  format: float
                  example: 416.67
                emi_date:
                  type: integer
                  example: 15
                insurer:
                  type: string
                  example: "ABC Insurance"
                insurance_policy_number:
                  type: string
                  example: "POL123456"
                insurance_start_date:
                  type: string
                  format: date
                  example: "2023-01-15"
                insurance_expiry_date:
                  type: string
                  format: date
                  example: "2024-01-14"
                pollution_certificate_expiry_date:
                  type: string
                  format: date
                  example: "2024-06-30"
                fitness_expiry_date:
                  type: string
                  format: date
                  example: "2024-06-30"
                passing_expiry_date:
                  type: string
                  format: date
                  example: "2024-06-30"
                mv_tax_expiry_date:
                  type: string
                  format: date
                  example: "2024-03-31"
                auth_date_expiry:
                  type: string
                  format: date
                  example: "2024-12-31"
                odometer_reading:
                  type: integer
                  example: 15000
                notes:
                  type: string
                  example: "Brand new vehicle"
                feature_image:
                  type: string
                  format: binary
                  description: "Vehicle feature image"
      responses:
        201:
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Vehicle created successfully"
                  data:
                    type: object
                    properties:
                      vehicleId:
                        type: integer
                        example: 1
        400:
          description: Validation error
        500:
          description: Internal server error

    get:
      tags: [Vehicles]
      summary: Get all vehicles with pagination and search
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            example: 1
        - name: search
          in: query
          required: false
          description: Search in chassis_number, make, title, model, fuel_type, registration_number
          schema:
            type: string
            example: "Toyota Camry"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by
          schema:
            type: string
            enum: [id, make, model, year, created_at, updated_at]
            default: id
            example: id
        - name: sort_order
          in: query
          required: false
          description: Sort order direction
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
            example: DESC
      responses:
        200:
          description: List of vehicles with pagination info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      vehicles:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 1
                            user_id:
                              type: integer
                              example: 1
                            vehicle_type_id:
                              type: integer
                              example: 1
                            vehicle_type_name:
                              type: string
                              example: "Sedan"
                            make:
                              type: string
                              example: "Toyota"
                            title:
                              type: string
                              example: "Toyota Camry 2023"
                            model:
                              type: string
                              example: "Camry"
                            year:
                              type: integer
                              example: 2023
                            chassis_number:
                              type: string
                              example: "CHS123456789"
                            fuel_type:
                              type: string
                              example: "Petrol"
                            registration_number:
                              type: string
                              example: "ABC123"
                            status:
                              type: string
                              example: "active"
                            feature_image:
                              type: string
                              nullable: true
                              example: "https://bucket.s3.region.amazonaws.com/vehicles/image.jpg"
                            created_at:
                              type: string
                              format: date-time
                            updated_at:
                              type: string
                              format: date-time
                      pagination:
                        type: object
                        properties:
                          current_page:
                            type: integer
                            example: 1
                          total_pages:
                            type: integer
                            example: 5
                          total_count:
                            type: integer
                            example: 48
                          has_next_page:
                            type: boolean
                            example: true
                          has_prev_page:
                            type: boolean
                            example: false
                          next_page:
                            type: integer
                            nullable: true
                            example: 2
                          prev_page:
                            type: integer
                            nullable: true
                            example: null
                          limit:
                            type: integer
                            example: 10
        400:
          description: Invalid parameters or User ID is required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "User ID is required"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Failed to fetch vehicles"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT